extracted parse-blueprint module
